+++ admin/controller/payment/duitku_ovo.php
@@ -34,6 +34,7 @@ class ControllerPaymentDuitkuOvo extends Controller {
       'text_edit',
       'entry_merchant',
       'entry_api_key',          
+      'entry_expired_period',          
       'entry_test',
       'entry_total',
       'entry_order_status',