+++ catalog/controller/payment/duitku_vacimb.php
@@ -43,10 +43,76 @@ class ControllerPaymentDuitkuVacimb extends Controller {
     //generate Signature
     $merchant_code = $this->config->get('duitku_vacimb_merchant');
     $api_key = $this->config->get('duitku_vacimb_api_key');
+	$expired = $this->config->get('duitku_vacimb_expired') != null ? $this->config->get('duitku_vacimb_expired') : 1440;
     $order_id = $this->session->data['order_id'];
     $def_curr = $this->config->get('config_currency');
     $order_total = $def_curr == 'IDR' ? $order_info['total'] : $this->currency->convert($order_info['total'], $order_info['currency_code'], 'IDR');
-    $signature = md5($merchant_code . $order_id . intval($order_total) . $api_key);    
+    
+	//itemDetails
+	$products = $this->cart->getProducts();
+	  
+	$item_details = array();
+
+    foreach ($products as $product) {
+      $item = array(
+        'price'    => (int)$product['price']*(int)$product['quantity'],
+        'quantity' => (int)$product['quantity'],
+        'name'     => substr($product['name'], 0, 49)
+      );
+      $item_details[] = $item;
+    }
+
+    if ($this->cart->hasShipping()) {
+      $shipping_data = $this->session->data['shipping_method'];
+      if (($this->config->get('config_customer_price') && $this->customer->isLogged()) || !$this->config->get('config_customer_price')) {
+        $shipping_data['cost'] = $this->tax->calculate(
+          $shipping_data['cost'],
+          $shipping_data['tax_class_id'],
+          $this->config->get('config_tax'));
+        }
+
+        $shipping_item = array(
+          'price' => (int)$shipping_data['cost'],
+          'quantity' => 1,
+          'name' => 'Shipping Fee'
+        );
+        $item_details[] = $shipping_item;
+	}
+
+	$amount_price = 0;
+	foreach ($item_details as $item) {
+		$amount_price += $item['price'];
+	}
+
+	if ($amount_price != $order_total) {
+		$coupon_item = array(
+			'price'    => (int)$order_total - (int)$amount_price,
+			'quantity' => 1,
+			'name'     => 'Coupon'
+		  );
+		$item_details[] = $coupon_item;
+	}
+	
+	$billing_address = array(
+	  'firstName' => $order_info['payment_firstname'],
+	  'lastName' => $order_info['payment_lastname'],
+	  'address' => $order_info['payment_address_1'],
+	  'city' => $order_info['payment_city'],
+	  'postalCode' => $order_info['payment_postcode'],
+	  'phone' => $order_info['telephone'],
+	  'countryCode' => "ID"
+	);
+	
+	$customerDetails = array(
+		'firstName' => $order_info['payment_firstname'],
+		'lastName' => $order_info['payment_lastname'],
+		'email' => $order_info['email'],
+		'phoneNumber' => $order_info['telephone'],
+		'billingAddress' => $billing_address,
+		'shippingAddress' => $billing_address
+	);
+	
+	$signature = md5($merchant_code . $order_id . intval($order_total) . $api_key);    
 
     // Prepare Parameters
     $params = array(
@@ -55,19 +121,25 @@ class ControllerPaymentDuitkuVacimb extends Controller {
           'paymentMethod' => "B1",
           'merchantOrderId' => $order_id,
           'productDetails' => $this->config->get('config_name') . ' Order : #' . $order_id,
-          'additionalParam' => '',
-          'merchantUserInfo' => $order_info['payment_firstname'] . " " . $order_info['payment_lastname'],
-          'signature' => $signature,          
+          'additionalParam' => $order_info['payment_firstname'] . " " . $order_info['payment_lastname'],
+          'merchantUserInfo' => $order_info['email'],
+		  'customerVaName' => $order_info['email'],
+		  'email' => $order_info['email'],
+          'phoneNumber' => $order_info['telephone'],
+          'signature' => $signature,
+		  'expiryPeriod' => $expired,       
           'returnUrl' => $this->url->link('payment/duitku_vacimb/landing_redir'),
           'callbackUrl' => $this->url->link('payment/duitku_vacimb/payment_notification'),
-    );           
+		  'customerDetail' => $customerDetails,
+		  'itemDetails' => $item_details,
+    );          
 
    /* Duitku_Config::$isProduction =
         $this->config->get('duitku_environment') == 'production'
         ? true : false;   */
 
     try {     
-      $redirUrl = Duitku_VtWeb::getRedirectionUrl($this->config->get('duitku_vacimb_endpoint'), $params);       
+      $redirUrl = DuitkuCore_Web::getRedirectionUrl($this->config->get('duitku_vacimb_endpoint'), $params);       
        $this->redirect($redirUrl);
     }
     catch (Exception $e) {
@@ -81,7 +153,7 @@ class ControllerPaymentDuitkuVacimb extends Controller {
    * Landing page when payment is finished or failure or customer pressed "back" button
    * The Cart is cleared here, so make sure customer reach this page to ensure the cart is emptied when payment succeed
    * payment finish/unfinish/error url :
-   * http://[your shop’s homepage]/index.php?route=payment/veritrans/payment_notification
+   * http://[your shop’s homepage]/index.php?route=payment/duitku_vacimb/payment_notification
    */
   public function landing_redir() {    
         
@@ -166,7 +238,7 @@ class ControllerPaymentDuitkuVacimb extends Controller {
 
     //check if order id is in the database
     if ($order_info) {        
-        if ($status == '00' && Duitku_VtWeb::validateTransaction($endpoint, $merchant_code, $order_id, $reference, $api_key)) {
+        if ($status == '00' && DuitkuCore_Web::validateTransaction($endpoint, $merchant_code, $order_id, $reference, $api_key)) {
           $order_status_id = $this->config->get('duitku_vacimb_success_mapping');          
         } else {
           $order_status_id = $this->config->get('duitku_vacimb_failure_mapping');