+++ changelog.md
@@ -16,4 +16,12 @@
 - add channels to Mandiri Virtual Account
 
 
+2.0
+-----
+- Upgrade Duitku API V2
+- add channels to ShopeePay
+- add channels to Indodana
 
+2.1
+-----
+- Improve parameter Duitku Expiry Period